Swipe right when negative speed

+++ autocat_enacter/Interaction.h
@@ -37,7 +37,7 @@ protected:
   bool _is_focussed = false;
   int _focus_x = 0;
   int _focus_y = 0;
-  int _focus_speed = 0;
+  float _focus_speed = 0.;  // Must be a float to multiply by the elapsed time
   int _clock = 0;
   int head_destination_angle = 0;
   unsigned long _duration1;
